Migrated activity logs api tests
