+++ package.json
@@ -36,6 +36,7 @@
     "test:integration-parallel": "jest --forceExit --maxWorkers=4 --config ./test/config/jest.config.integration.js",
     "test:organization": "jest --forceExit --maxWorkers=6 --config ./test/config/jest.config.organization.js",
     "test:search": "jest --forceExit --maxWorkers=6 --config ./test/config/jest.config.search.js",
+    "test:activity-logs": "jest --forceExit --maxWorkers=6 --config ./test/config/jest.config.activity-logs.js",
     "test:usermanagement": "jest --forceExit --maxWorkers=4 --config ./test/config/jest.config.usermanagement.js",
     "test:integration": "jest --forceExit --maxWorkers=1 --verbose --config ./test/config/jest.config.integration.js",
     "test:upload": "jest --forceExit --maxWorkers=1 --verbose --config ./test/config/jest.config.upload.js",
@@ -68,7 +69,7 @@
   },
   "devDependencies": {
     "@graphql-codegen/add": "^3.2.1",
-    "@graphql-codegen/cli": "^2.11.8",
+    "@graphql-codegen/cli": "^2.16.5",
     "@graphql-codegen/typescript": "2.7.3",
     "@graphql-codegen/typescript-resolvers": "2.7.3",
     "@graphql-codegen/import-types-preset": "^2.2.3",
