+++ test/functional-api/activity-logs/activity-log-params.ts
@@ -1,0 +1,25 @@
+import { TestUser } from '@test/utils';
+import { graphqlErrorWrapper } from '@test/utils/graphql.wrapper';
+import { getGraphqlClient } from '@test/utils/graphqlClient';
+
+export const getActivityLogOnCollaborationCodegen = async (
+  collaborationID: string,
+  limit: number,
+  userRole: TestUser = TestUser.GLOBAL_ADMIN
+) => {
+  const graphqlClient = getGraphqlClient();
+  const callback = (authToken: string | undefined) =>
+    graphqlClient.GetActivityLogOnCollaboration(
+      {
+        queryData: {
+          collaborationID,
+          limit,
+        },
+      },
+      {
+        authorization: `Bearer ${authToken}`,
+      }
+    );
+
+  return graphqlErrorWrapper(callback, userRole);
+};
