--- test/functional-api/integration/activity-log-collaboration/activity-logs-enum.ts
@@ -1,8 +1,0 @@
-export enum ActivityLogs {
-  CALLOUT_PUBLISHED = 'CALLOUT_PUBLISHED',
-  WHITEBOARD_CREATED = 'WHITEBOARD_CREATED',
-  POST_COMMENT = 'POST_COMMENT',
-  POST_CREATED = 'POST_CREATED',
-  DISCUSSION_COMMENT = 'DISCUSSION_COMMENT',
-  MEMBER_JOINED = 'MEMBER_JOINED',
-}
